Rename NamuiBundleList to NamuiBundleManifest according to changed structure
