--- namui-cli/src/util/get_namui_bundle_list/mod.rs
@@ -1,4 +1,0 @@
-mod get_namui_bundle_list;
-mod lexicon;
-mod token;
-pub use get_namui_bundle_list::get_namui_bundle_list;
