+++ namui-cli/src/util/mod.rs
@@ -5,8 +5,8 @@ mod print_build_result;
 pub use print_build_result::*;
 mod get_cli_root_path;
 pub use get_cli_root_path::*;
-mod get_namui_bundle_list;
-pub use get_namui_bundle_list::*;
+mod namui_bundle_manifest;
+pub use namui_bundle_manifest::*;
 mod get_electron_root_path;
 pub use get_electron_root_path::*;
 mod overwrite_hot_reload_script_with_empty_file;
