+++ namui-cli/src/util/namui_bundle_manifest/lexicon/mod.rs
@@ -1,8 +1,8 @@
 mod lexer;
-mod namui_bundle_list;
+mod namui_bundle_manifest;
 mod operation;
 mod path;
 pub use lexer::Lexer;
-pub use namui_bundle_list::NamuiBundleList;
+pub use namui_bundle_manifest::NamuiBundleManifest;
 use operation::*;
 use path::*;
