+++ namui-cli/src/util/namui_bundle_manifest/lexicon/operation.rs
